+++ tests/integration/cloudformation/resources/test_cloudformation_cdk.py
@@ -17,7 +17,8 @@ class TestCdkInit:
     def test_cdk_bootstrap(self, deploy_cfn_template, cfn_client, bootstrap_version):
         deploy_cfn_template(
             template_path=os.path.join(
-                os.path.dirname(__file__), f"../../templates/cdk_bootstrap_v{bootstrap_version}.yaml"
+                os.path.dirname(__file__),
+                f"../../templates/cdk_bootstrap_v{bootstrap_version}.yaml",
             )
         )
         init_stack_result = deploy_cfn_template(
