--- tests/integration/cloudformation/resources/test_cloudformation_dynamic_parameters.py
@@ -1,7 +1,0 @@
-import os
-
-import pytest as pytest
-
-from localstack.utils.common import short_uid
-
-
