+++ tests/integration/cloudformation/resources/test_cloudformation_s3.py
@@ -12,7 +12,9 @@ def test_bucketpolicy(
 ):
     bucket_name = f"ls-bucket-{short_uid()}"
     deploy_result = deploy_cfn_template(
-        template_path=os.path.join(os.path.dirname(__file__), "../../templates/s3_bucketpolicy.yaml"),
+        template_path=os.path.join(
+            os.path.dirname(__file__), "../../templates/s3_bucketpolicy.yaml"
+        ),
         parameters={"BucketName": bucket_name},
         template_mapping={"include_policy": True},
     )
@@ -23,7 +25,9 @@ def test_bucketpolicy(
         is_update=True,
         stack_name=deploy_result.stack_id,
         parameters={"BucketName": bucket_name},
-        template_path=os.path.join(os.path.dirname(__file__), "../../templates/s3_bucketpolicy.yaml"),
+        template_path=os.path.join(
+            os.path.dirname(__file__), "../../templates/s3_bucketpolicy.yaml"
+        ),
         template_mapping={"include_policy": False},
     )
     with pytest.raises(Exception) as err:
