+++ tests/integration/cloudformation/test_cloudformation_stacks.py
@@ -4,71 +4,9 @@ import os
 import pytest
 import yaml
 
-from localstack.testing.aws.cloudformation_utils import load_template_file, load_template_raw
-from localstack.utils.aws import aws_stack
+from localstack.testing.aws.cloudformation_utils import load_template_file
 from localstack.utils.files import load_file
 from localstack.utils.strings import short_uid
-from localstack.utils.sync import retry, wait_until
-
-
-def test_create_stack_with_ssm_parameters(cfn_client, ssm_client, sns_client, deploy_cfn_template):
-    parameter_name = f"ls-param-{short_uid()}"
-    parameter_value = f"ls-param-value-{short_uid()}"
-    parameter_logical_id = "parameter123"
-    ssm_client.put_parameter(Name=parameter_name, Value=parameter_value, Type="String")
-    stack = deploy_cfn_template(
-        template_path=os.path.join(
-            os.path.dirname(__file__), "../templates/dynamicparameter_ssm_string.yaml"
-        ),
-        template_mapping={"parameter_name": parameter_name},
-    )
-
-    stack_description = cfn_client.describe_stacks(StackName=stack.stack_name)["Stacks"][0]
-    assert stack_description is not None
-    assert stack_description["Parameters"][0]["ParameterKey"] == parameter_logical_id
-    assert stack_description["Parameters"][0]["ParameterValue"] == parameter_name
-    assert stack_description["Parameters"][0]["ResolvedValue"] == parameter_value
-
-    topics = sns_client.list_topics()
-    topic_arns = [t["TopicArn"] for t in topics["Topics"]]
-    assert any(parameter_value in t for t in topic_arns)
-
-
-def test_list_stack_resources_for_removed_resource(cfn_client, deploy_cfn_template):
-    template_path = os.path.join(os.path.dirname(__file__), "../templates/eventbridge_policy.yaml")
-    event_bus_name = f"bus-{short_uid()}"
-    stack = deploy_cfn_template(
-        template_path=template_path,
-        parameters={"EventBusName": event_bus_name},
-    )
-
-    resources = cfn_client.list_stack_resources(StackName=stack.stack_name)[
-        "StackResourceSummaries"
-    ]
-    resources_before = len(resources)
-    assert resources_before == 3
-    statuses = set([res["ResourceStatus"] for res in resources])
-    assert statuses == {"CREATE_COMPLETE"}
-
-    # remove one resource from the template, then update stack (via change set)
-    template_dict = yaml.safe_load(open(template_path))
-    template_dict["Resources"].pop("eventPolicy2")
-    template2 = yaml.dump(template_dict)
-
-    deploy_cfn_template(
-        stack_name=stack.stack_name,
-        is_update=True,
-        template=template2,
-        parameters={"EventBusName": event_bus_name},
-    )
-
-    # get list of stack resources, again - make sure that deleted resource is not contained in result
-    resources = cfn_client.list_stack_resources(StackName=stack.stack_name)[
-        "StackResourceSummaries"
-    ]
-    assert len(resources) == resources_before - 1
-    statuses = set([res["ResourceStatus"] for res in resources])
-    assert statuses == {"UPDATE_COMPLETE"}
 
 
 @pytest.mark.aws_validated
@@ -115,57 +53,6 @@ Resources:
     deployed = deploy_cfn_template(template=template)
     response = cfn_client.describe_stacks(StackName=deployed.stack_id)["Stacks"][0]
     snapshot.match("describe_stack", response)
-
-
-@pytest.mark.aws_validated
-def test_import_values_across_stacks(deploy_cfn_template, s3_client):
-    export_name = f"b-{short_uid()}"
-
-    # create stack #1
-    template1 = """
-Parameters:
-  BucketExportName:
-    Type: String
-Resources:
-  Bucket1:
-    Type: AWS::S3::Bucket
-    Properties: {}
-Outputs:
-  BucketName1:
-    Value: !Ref Bucket1
-    Export:
-      Name: !Ref BucketExportName
-    """
-    result = deploy_cfn_template(template=template1, parameters={"BucketExportName": export_name})
-    bucket_name1 = result.outputs.get("BucketName1")
-    assert bucket_name1
-
-    # create stack #2
-    template2 = """
-Parameters:
-  BucketExportName:
-    Type: String
-Resources:
-  Bucket2:
-    Type: AWS::S3::Bucket
-    Properties:
-      Tags:
-        - Key: test
-          Value: !ImportValue
-            'Fn::Sub': '${BucketExportName}'
-Outputs:
-  BucketName2:
-    Value: !Ref Bucket2
-    """
-    result = deploy_cfn_template(template=template2, parameters={"BucketExportName": export_name})
-    bucket_name2 = result.outputs.get("BucketName2")
-    assert bucket_name2
-
-    # assert that correct bucket tags have been created
-    tagging = s3_client.get_bucket_tagging(Bucket=bucket_name2)
-    test_tag = [tag for tag in tagging["TagSet"] if tag["Key"] == "test"]
-    assert test_tag
-    assert test_tag[0]["Value"] == bucket_name1
 
 
 @pytest.mark.aws_validated
@@ -237,33 +124,41 @@ def test_stack_update_resources(
     snapshot.match("stack_resources", resources)
 
 
-def test_nested_stack(s3_client, cfn_client, deploy_cfn_template, s3_create_bucket):
-    # upload template to S3
-    artifacts_bucket = f"cf-artifacts-{short_uid()}"
-    artifacts_path = "stack.yaml"
-    s3_create_bucket(Bucket=artifacts_bucket, ACL="public-read")
-    s3_client.put_object(
-        Bucket=artifacts_bucket,
-        Key=artifacts_path,
-        Body=load_file(os.path.join(os.path.dirname(__file__), "../templates/template5.yaml")),
+def test_list_stack_resources_for_removed_resource(cfn_client, deploy_cfn_template):
+    template_path = os.path.join(os.path.dirname(__file__), "../templates/eventbridge_policy.yaml")
+    event_bus_name = f"bus-{short_uid()}"
+    stack = deploy_cfn_template(
+        template_path=template_path,
+        parameters={"EventBusName": event_bus_name},
     )
 
-    # deploy template
-    param_value = short_uid()
-    stack_bucket_name = f"test-{param_value}"  # this is the bucket name generated by template5
+    resources = cfn_client.list_stack_resources(StackName=stack.stack_name)[
+        "StackResourceSummaries"
+    ]
+    resources_before = len(resources)
+    assert resources_before == 3
+    statuses = set([res["ResourceStatus"] for res in resources])
+    assert statuses == {"CREATE_COMPLETE"}
+
+    # remove one resource from the template, then update stack (via change set)
+    template_dict = yaml.safe_load(open(template_path))
+    template_dict["Resources"].pop("eventPolicy2")
+    template2 = yaml.dump(template_dict)
 
     deploy_cfn_template(
-        template=load_file(os.path.join(os.path.dirname(__file__), "../templates/template6.yaml"))
-        % (artifacts_bucket, artifacts_path),
-        parameters={"GlobalParam": param_value},
+        stack_name=stack.stack_name,
+        is_update=True,
+        template=template2,
+        parameters={"EventBusName": event_bus_name},
     )
 
-    # assert that nested resources have been created
-    def assert_bucket_exists():
-        response = s3_client.head_bucket(Bucket=stack_bucket_name)
-        assert 200 == response["ResponseMetadata"]["HTTPStatusCode"]
-
-    retry(assert_bucket_exists)
+    # get list of stack resources, again - make sure that deleted resource is not contained in result
+    resources = cfn_client.list_stack_resources(StackName=stack.stack_name)[
+        "StackResourceSummaries"
+    ]
+    assert len(resources) == resources_before - 1
+    statuses = set([res["ResourceStatus"] for res in resources])
+    assert statuses == {"UPDATE_COMPLETE"}
 
 
 def test_update_stack_with_same_template(cfn_client, deploy_cfn_template):
@@ -277,7 +172,3 @@ def test_update_stack_with_same_template(cfn_client, deploy_cfn_template):
     error_message = str(ctx.value)
     assert "UpdateStack" in error_message
     assert "No updates are to be performed." in error_message
-
-
-
-
